fix(subgrid): cast coarse-source scatter to dest dtype too (#485 follow-up)
_inject_coarse_sources carries the identical float64->float32 x64 leak as
_inject_fine_sources (fixed in 64d96f3): src_vals_c is built via
jnp.array(waveform) (jit_runner.py ~2627), which tracks x64, while
ex/ey/ez_arr stay float32 regardless (init_state's field_dtype default).
Same FutureWarning class, same fix -- cast the source value to the
destination array's own dtype at each of the three coarse scatter sites.

Coverage: the coarse path only fires when src_meta_c is non-empty, which
needs a soft source (Simulation.add_source, impedance=0 -- add_port/lumped
ports never populate sources_c) plus inject_sources_on_coarse_shadow
enabled. Both are cheap (no special coarse-fed configuration needed), so
this is a live-gated test, not cast-by-symmetry: added
_coarse_shadow_source_subgrid_sim + two tests
(test_coarse_source_injection_no_dtype_futurewarning_under_x64,
test_coarse_source_injection_default_precision_still_float32) to
tests/test_subgrid_source_injection_dtype.py, verified both directions
locally:
  * pre-fix: FutureWarning raised at jit_runner.py:1776 inside
    _inject_coarse_sources (only the coarse test failed; fine tests still
    passed -- 1 failed, 3 passed).
  * post-fix: all 4 tests in the file pass, no FutureWarning.

Byte-identity at default precision (x64 off) re-verified for the coarse
fixture: time_series bit-identical (np.array_equal True, max abs diff 0.0,
dtype float32) pre-fix vs post-fix, via a patch-apply/revert cycle (not
git stash, per the shared-tree rule).

Full subgrid suite (14 files + this test file): 116 passed before this
commit -> 118 passed after (2 new), same 4 deselected / 2 xfailed /
warnings -- no regressions.

R3: memory=none-after-grep (rfx-known-issues.md, index.md; same precedent
as 64d96f3, consistent not contradicted) | R2-attempts=0 (single clean
attempt; coarse-path exercisability was investigated once via the
sources_c/impedance==0.0 code path in rfx/runners/subgridded.py, not a
non-closing retry) | falsifier=rerun
test_subgrid_source_injection_dtype.py::test_coarse_source_injection_no_dtype_futurewarning_under_x64
with warnings.simplefilter("error", FutureWarning) under scoped x64

+++ rfx/subgridding/jit_runner.py
@@ -1771,13 +1771,18 @@ def _make_step_fn(ctx):
             return ex_arr, ey_arr, ez_arr
 
         def _inject_coarse_sources(ex_arr, ey_arr, ez_arr):
+            # Same leak as _inject_fine_sources above (issue #485): src_vals_c
+            # is built from jnp.array(waveform) (jit_runner.py ~2627), which
+            # tracks x64, while ex/ey/ez_arr stay float32 regardless
+            # (init_state's field_dtype default). Cast explicitly to the
+            # destination array's own dtype for the same reason.
             for idx_s, (si, sj, sk, sc) in enumerate(src_meta_c):
                 if sc == "ez":
-                    ez_arr = ez_arr.at[si, sj, sk].add(src_vals_c[idx_s])
+                    ez_arr = ez_arr.at[si, sj, sk].add(src_vals_c[idx_s].astype(ez_arr.dtype))
                 elif sc == "ex":
-                    ex_arr = ex_arr.at[si, sj, sk].add(src_vals_c[idx_s])
+                    ex_arr = ex_arr.at[si, sj, sk].add(src_vals_c[idx_s].astype(ex_arr.dtype))
                 elif sc == "ey":
-                    ey_arr = ey_arr.at[si, sj, sk].add(src_vals_c[idx_s])
+                    ey_arr = ey_arr.at[si, sj, sk].add(src_vals_c[idx_s].astype(ey_arr.dtype))
             return ex_arr, ey_arr, ez_arr
 
         if sync_coarse_shadow_from_fine and is_full_xy_z_slab:
